+++ renderer/src/pages/Git/components/SSHKeyFormItemLabel/index.tsx
@@ -5,13 +5,13 @@ const { Tooltip } = Balloon;
 export default () => {
   return (
     <span>
-      使用此配置的目录
+      SSH 公钥
       <Tooltip
         trigger={<Icon type="prompt" style={{ marginLeft: 4 }} />}
         align="t"
         delay={200}
       >
-        该目录下的 Git 项目都将使用此配置，详细说明查看<a href="https://git-scm.com/docs/git-config#_conditional_includes" target="__blank">文档</a>
+        添加 SSH 公钥教程，请查看<a href="https://appworks.site/pack/basic/toolkit.html" target="__blank">文档</a>
       </Tooltip>
     </span>
   );
